+++ CHANGELOG.md
@@ -159,6 +159,7 @@ function:
 - console: enable searching tables within a schema
 - console: fixed the ability to create updated_at and created_at in the modify page (#8239)
 - console: disable search indexing with HTML meta tag
+- console: add support for setting comments on the custom root fields of tables/views
 - cli: fix inherited roles metadata not being updated when dropping all roles (#7872)
 - cli: add support for customization field in sources metadata (#8292)
 - ci: ubuntu and centos flavoured graphql-engine images are now available